Bug fixed:
・履歴に存在するファイルは削除しない設定が正しく動作しない問題を解消。

+++ Implem.Pleasanter/Libraries/DataTypes/Attachment.cs
@@ -91,19 +91,20 @@ namespace Implem.Pleasanter.Libraries.DataTypes
             Context context,
             SiteSettings ss,
             Column column,
-            long referenceId)
+            long referenceId,
+            bool verUp)
         {
             var path = Path.Combine(
                 Directories.BinaryStorage(),
                 "Attachments",
                 Guid);
-            if (ss == null
-                || column.NotDeleteExistHistory == false
-                || !ExistsHistory(
-                    context: context,
-                    ss: ss,
-        　　　　    column: column,
-        　　　　    referenceId: referenceId))
+            if (column?.NotDeleteExistHistory != true
+                || (verUp == false
+                    && !ExistsHistory(
+                        context: context,
+                        ss: ss,
+                        column: column,
+                        referenceId: referenceId)))
             {
                 if (System.IO.File.Exists(path))
                 {
@@ -117,7 +118,8 @@ namespace Implem.Pleasanter.Libraries.DataTypes
             SiteSettings ss,
             Column column,
             List<SqlStatement> statements,
-            long referenceId)
+            long referenceId,
+            bool verUp)
         {
             if (Added == true)
             {
@@ -140,12 +142,13 @@ namespace Implem.Pleasanter.Libraries.DataTypes
             }
             else if (Deleted == true && !Overwritten.HasValue)
             {
-                if (column.NotDeleteExistHistory == false
-                    || !ExistsHistory(
-                        context: context,
-                        ss: ss,
-                        column: column,
-                        referenceId: referenceId))
+                if (column?.NotDeleteExistHistory != true
+                    || (verUp == false
+                        && !ExistsHistory(
+                            context: context,
+                            ss: ss,
+                            column: column,
+                            referenceId: referenceId)))
                 {
                     statements.Add(Rds.DeleteBinaries(
                         factory: context,
@@ -174,6 +177,10 @@ namespace Implem.Pleasanter.Libraries.DataTypes
             Column column,
             long referenceId)
         {
+            if (ss == null)
+            {
+                return false;
+            }
             switch (ss.ReferenceType)
             {
                 case "Issues":
@@ -190,8 +197,7 @@ namespace Implem.Pleasanter.Libraries.DataTypes
                                     columnBrackets: column.ColumnName.ToSingleArray(),
                                     value: $"%\"Guid\":\"{Guid}\"%",
                                     name: Strings.NewGuid(),
-                                    _operator: context.Sqls.LikeWithEscape),
-                            top: 1)) == 1;
+                                    _operator: context.Sqls.LikeWithEscape))) >= 1;
                 case "Results":
                     return Rds.ExecuteScalar_int(
                         context: context,
@@ -206,8 +212,7 @@ namespace Implem.Pleasanter.Libraries.DataTypes
                                     columnBrackets: column.ColumnName.ToSingleArray(),
                                     value: $"%\"Guid\":\"{Guid}\"%",
                                     name: Strings.NewGuid(),
-                                    _operator: context.Sqls.LikeWithEscape),
-                            top: 1)) == 1;
+                                    _operator: context.Sqls.LikeWithEscape))) >= 1;
                 default:
                     return false;
             }
